Scaladoc tests: share compile-at-test-time helpers
